raftstore: extract worker::metrics
