+++ components/raftstore/src/store/worker/mod.rs
@@ -1,1 +1,2 @@
 pub mod cleanup_sst;
+pub mod metrics;
